+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.2 (unreleased)
 
-* New Feature: Create initial snapshot of VM after clean-up
+* New Feature: Create initial snapshot of VM after clean-up [GH-21]
+* Enable hpet by default [GH-3]
 
 ## 0.0.1 (April 19, 2021)